+++ src/tracing/instrumentation/mysql.js
@@ -12,6 +12,7 @@ var isActive = false;
 exports.init = function() {
   requireHook.on('mysql', instrumentMysql);
   requireHook.on('mysql2', instrumentMysql2);
+  requireHook.on('mysql2/promise', instrumentMysql2WithPromises);
 };
 
 
@@ -26,6 +27,11 @@ function instrumentMysql2(mysql) {
   mysql.Pool && instrumentPool(mysql.Pool.prototype);
 }
 
+function instrumentMysql2WithPromises(mysql) {
+  // Currently only pooled connection will be instrumented
+  instrumentPoolWithPromises(mysql);
+}
+
 
 function instrumentPool(Pool) {
   shimmer.wrap(Pool, 'query', shimQuery);
@@ -37,6 +43,15 @@ function instrumentConnection(Connection) {
   shimmer.wrap(Connection, 'query', shimQuery);
 }
 
+function instrumentPoolWithPromises(mysql) {
+  shimmer.wrap(mysql, 'createPool', function(original) {
+    return function() {
+      var Pool = original.apply(this, arguments);
+      shimmer.wrap(Object.getPrototypeOf(Pool), 'getConnection', shimPromiseConnection);
+      return Pool;
+    };
+  });
+}
 
 function shimQuery(original) {
   return function() {
@@ -47,7 +62,16 @@ function shimQuery(original) {
   };
 }
 
-function instrumentedQuery(ctx, originalQuery, statementOrOpts, valuesOrCallback, optCallback) {
+function shimPromiseQuery(originalQuery) {
+  return function() {
+    if (isActive && cls.isTracing()) {
+      return instrumentedQuery(this, originalQuery, arguments[0], arguments[1], null, true);
+    }
+    return originalQuery.apply(this, arguments);
+  };
+}
+
+function instrumentedQuery(ctx, originalQuery, statementOrOpts, valuesOrCallback, optCallback, isPromiseImpl) {
   var argsForOriginalQuery = [statementOrOpts, valuesOrCallback];
   if (typeof optCallback !== 'undefined') {
     argsForOriginalQuery.push(optCallback);
@@ -62,34 +86,56 @@ function instrumentedQuery(ctx, originalQuery, statementOrOpts, valuesOrCallback
   var port;
   var user;
   var db;
-  if (ctx.config) {
-    if (ctx.config.connectionConfig) {
-      host = ctx.config.connectionConfig.host;
-      port = ctx.config.connectionConfig.port;
-      user = ctx.config.connectionConfig.user;
-      db = ctx.config.connectionConfig.database;
+
+  // if ctx.connection is defined, we are in a PromiseConnection context
+  var config = ctx.connection != null ? ctx.connection.config : ctx.config;
+  if (config) {
+    if (config.connectionConfig) {
+      host = config.connectionConfig.host;
+      port = config.connectionConfig.port;
+      user = config.connectionConfig.user;
+      db = config.connectionConfig.database;
     } else {
-      host = ctx.config.host;
-      port = ctx.config.port;
-      user = ctx.config.user;
-      db = ctx.config.database;
+      host = config.host;
+      port = config.port;
+      user = config.user;
+      db = config.database;
     }
   }
 
   var span = cls.startSpan('mysql');
-  span.b = { s: 1 };
+  span.b = {s: 1};
   span.stack = tracingUtil.getStackTrace(instrumentedQuery);
   span.data = {
-      mysql: {
-        stmt: tracingUtil.shortenDatabaseStatement(typeof statementOrOpts === 'string' ? statementOrOpts :
-          statementOrOpts.sql),
-        host: host,
-        port: port,
-        user: user,
-        db: db
-      }
-    };
+    mysql: {
+      stmt: tracingUtil.shortenDatabaseStatement(typeof statementOrOpts === 'string' ? statementOrOpts :
+        statementOrOpts.sql),
+      host: host,
+      port: port,
+      user: user,
+      db: db
+    }
+  };
 
+  if (isPromiseImpl) {
+    var resultPromise = originalQuery.apply(ctx, argsForOriginalQuery);
+
+    resultPromise.then(function(result) {
+      span.d = Date.now() - span.ts;
+      transmission.addSpan(span);
+      return result;
+    }).catch(function(error) {
+        span.ec = 1;
+        span.error = true;
+        span.data.mysql.error = tracingUtil.getErrorDetails(error);
+
+        span.d = Date.now() - span.ts;
+        transmission.addSpan(span);
+        return error;
+      });
+    return resultPromise;
+  }
+  // no promise, continue with standard instrumentation
   var originalCallback = argsForOriginalQuery[argsForOriginalQuery.length - 1];
   argsForOriginalQuery[argsForOriginalQuery.length - 1] = function onQueryResult(error) {
     if (error) {
@@ -113,6 +159,17 @@ function instrumentedQuery(ctx, originalQuery, statementOrOpts, valuesOrCallback
 function shimGetConnection(original) {
   return function(cb) {
     return original.call(this, cls.ns.bind(cb));
+  };
+}
+
+function shimPromiseConnection(original) {
+  return function getConnection() {
+    return original.apply(this, arguments)
+      .then(function(connection) {
+        shimmer.wrap(connection, 'query', shimPromiseQuery);
+
+        return connection;
+      });
   };
 }